Fixes bugs and refactoring
