+++ ts/screens/profile/SecurityScreen.tsx
@@ -2,26 +2,22 @@ import React, { FC, useEffect, useState } from "react";
 import { List } from "native-base";
 import { connect } from "react-redux";
 import { Dispatch } from "redux";
-import FingerprintScanner, {
-  AuthenticateAndroid,
-  AuthenticateIOS
-} from "react-native-fingerprint-scanner";
-import { Platform } from "react-native";
 import I18n from "../../i18n";
 import { GlobalState } from "../../store/reducers/types";
-import { getFingerprintSettings } from "../../sagas/startup/checkAcknowledgedFingerprintSaga";
 import TopScreenComponent from "../../components/screens/TopScreenComponent";
 import { ContextualHelpPropsMarkdown } from "../../components/screens/BaseScreenComponent";
 import ListItemComponent from "../../components/screens/ListItemComponent";
 import { updatePin } from "../../store/actions/pinset";
 import { identificationRequest } from "../../store/actions/identification";
 import { shufflePinPadOnPayment } from "../../config";
-import { authenticateConfig } from "../../utils/biometric";
+import {
+  biometricAuthenticationRequest,
+  getFingerprintSettings
+} from "../../utils/biometric";
 import { showToast } from "../../utils/showToast";
 import { preferenceFingerprintIsEnabledSaveSuccess } from "../../store/actions/persistedPreferences";
 import { useScreenReaderEnabled } from "../../utils/accessibility";
 import ScreenContent from "../../components/screens/ScreenContent";
-import { mixpanelTrack } from "../../mixpanel";
 
 const contextualHelpMarkdown: ContextualHelpPropsMarkdown = {
   title: "profile.preferences.contextualHelpTitle",
@@ -60,37 +56,16 @@ const SecurityScreen: FC<Props> = ({
       return;
     }
     // if user asks to disable biometric recnognition is required to proceed
-    FingerprintScanner.authenticate(
-      Platform.select({
-        ios: {
-          description: I18n.t(
-            "identification.biometric.popup.sensorDescription"
-          ),
-          fallbackEnabled: true
-        } as AuthenticateIOS,
-        default: {
-          title: authenticateConfig.title,
-          description: I18n.t(
-            "identification.biometric.popup.sensorDescription"
-          ),
-          cancelButton: I18n.t("global.buttons.cancel")
-        } as AuthenticateAndroid
-      })
-    )
-      .then(() => {
-        setFingerprintPreference(biometricPreference);
-        void FingerprintScanner.release();
-      })
-      .catch(e => {
-        void mixpanelTrack("BIOMETRIC_ERROR", { error: e });
+    biometricAuthenticationRequest(
+      () => setFingerprintPreference(biometricPreference),
+      _ =>
         showToast(
           I18n.t(
             "profile.security.list.biometric_recognition.needed_to_disable"
           ),
           "danger"
-        );
-        void FingerprintScanner.release();
-      });
+        )
+    );
   };
 
   return (
